Login System Alpha:
Added login system code in login/ directory.  Session management
will send you back to the login page and correctly destroy the session
if your session variables are not all perfect.  Session management code
was added tothe known list pages, but not the detail pages.  That's a
todo.

You can now see the user variables dynamically in the nav bar from the
list pages because of these additions to the codebase.